speed up guess_dates a little
It turns out that the most time spent in guess_dates was searching
for the first correct date since the data frame was being converted to a
matrix and then back again a couple of times due to the subsetting.

Now it's Thibaut's parsers that are responsible for the slowdown.